Extract CSS class merging logic from `Theme` into `CssClassMerger`.
